pyproject.toml
[tool.poetry]
name = "pyneapple"
version = "0.1.0"
description = ""
authors = ["siddhant0519"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.0"
geopandas = "^0.13.2"
pandas = "^2.0.3"
libpysal = "^4.7.0"
matplotlib = "^3.7.1"
jpype1 = "^1.4.1"
fastapi = "^0.98.0"
uvicorn = "^0.22.0"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
